break clean_data.py into two scripts
